% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT Functions.R
\name{cwtBoot}
\alias{cwtBoot}
\title{Bootstrapped CWT Confidence Intervals}
\usage{
cwtBoot(x, method = "parametric", nreps = 1000)
}
\arguments{
\item{x}{default cwt input}

\item{method}{The method to calculate bootstrapped intervals can be either "parametric" or
"geiger". Using parametric bootstrapping follows a deterministic approach; using Geiger's
method follows the Geiger (1990) approach to bootstrapping}

\item{nreps}{the number of fish sampled in the fishery}
}
\description{
Create a bootstrap confidence interval
}
\examples{
cwtBoot()
}
\keyword{cwt}
