% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT Functions.R
\name{cwtEst}
\alias{cwtEst}
\title{CWT Point Estimation}
\usage{
cwtEst(N, n, lambda, m, theta)
}
\arguments{
\item{N}{Number of fish harvested in the fishery, optionally including a vector with a length of 2, of
1) Number of fish harvested in fishery and 2) the associated variance. If entered as a single number, the
variance will default to NA.}

\item{n}{Number of fish sampled in the fishery (inspected for marks)}

\item{lambda}{Vector of length four containing 1) the number of heads with CWT shipped to the taglab, A1,
2) the number of heads received by the taglab, A2, 3) the number of tags detected at the taglab, M1, and
4) the number of tags successfully decoded, M2.}

\item{m}{Number of marked fish in the sample}

\item{theta}{the proportion of stock that is marked, optionally can be a vector of length two
containing 1) the proportion of stock that is marked and
2) if applicable, the associated variance of 1/theta, else NA}
}
\description{
This function estimates the point estimate for
}
\examples{
Point estimates for the first example from Geiger (1990)
cwtEst(N=c(1000,NA), n=200, lambda=c(1,1,1,1), m=10, theta=c(0.5,NA))

Alternately coded as:
cwtEst(N=1000, n=200, lambda=c(1,1,1,1), m=10, theta=0.5)
}
\keyword{cwt}
