% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwtr-package.R
\docType{package}
\name{cwtr-package}
\alias{cwtr}
\alias{cwtr-package}
\title{cwtr: CWT Helper Functions}
\description{
The premier package for calculating CWT estimates. CWT calculations of a single cohort can be found using this package.
}
\keyword{internal}
