% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT Functions.R
\name{plot.cwt}
\alias{plot.cwt}
\title{Plot CWT class data}
\usage{
\method{plot}{cwt}(x, parameters = c("r", "t", "q", "m"), type = "histogram")
}
\arguments{
\item{x}{default cwt input}

\item{parameters}{Vector of length 4 of "r", "t", "q", and "m"}

\item{type}{Plot type}
}
\description{
Four panel plot summary
}
\examples{
plot.cwt()
}
\keyword{cwt}
