% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT Functions.R
\name{print.cwt}
\alias{print.cwt}
\title{Print the CWT Confidence Interval Summary}
\usage{
\method{print}{cwt}(x, digits = 3)
}
\arguments{
\item{x}{Default cwt input}

\item{digits}{Number of digits to display, default=3}
}
\description{
Create a bootstrap confidence interval
}
\examples{
print.cwt()
}
\keyword{cwt}
