% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT Functions.R
\name{summary.cwt}
\alias{summary.cwt}
\title{CWT summary}
\usage{
\method{summary}{cwt}(x, alpha = 0.1, quiet = FALSE, ...)
}
\arguments{
\item{x}{default cwt input}

\item{alpha}{Alpha level for confidence, default=0.1}

\item{quiet}{Silence verbose notes, default = FALSE}
}
\description{
Display a summary of the
}
\examples{
summary.cwt()
}
\keyword{cwt}
