% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwtr-package.R
\docType{package}
\name{cwtr-package}
\alias{cwtr}
\alias{cwtr-package}
\title{cwtr: CWT Helper Functions}
\description{
Package 'cwtr' provides functions for calculating CWT estimates and bootstrap intervals. Simple CWT calculations of a single cohort can be found using the easy-to-use functions.
}
\author{
\strong{Maintainer}: Randy Peterson \email{randy.peterson@alaska.gov}

Authors:
\itemize{
  \item Randy Peterson \email{randy.peterson@alaska.gov}
  \item Justin Priest \email{justin.priest@alaska.gov}
}

}
\keyword{internal}
