% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT Functions.R
\name{print.summary.cwt}
\alias{print.summary.cwt}
\title{Print CWT summary}
\usage{
\method{print}{summary.cwt}(x, digits = 3)
}
\arguments{
\item{x}{Default cwt input}

\item{digits}{Number of digits to display, default=3}
}
\description{
Display a summary of the
}
\examples{
print.summary.cwt()
}
\keyword{cwt}
